group_17/pyproject.toml
[project]
name = "devops-incident-suite2"
version = "0.1.0"
description = "Multi-agent DevOps Incident Analysis Suite powered by Claude and LangGraph"
requires-python = ">=3.10"
dependencies = [
    # AI / Orchestration
    "anthropic",
    "langgraph",
    "langchain-core",
    # Web UI
    "streamlit",
    # REST API
    "fastapi",
    "uvicorn[standard]",
    "python-multipart",
    # Integrations
    "slack-sdk",
    "jira",
    # Utilities
    "python-dotenv",
]

[dependency-groups]
dev = [
    "ruff",
    "pytest",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
